refactor(crud files): refactor crud functions to have a table parameter
refactor crud functions to have a table parameter

The Secrets table will be passed into the crud functions rather than being hard coded into the function.

This is one refactor that will aid in being able to pass IAM auth tokens to the embrasure-server to authenticate non-admin users and allow them to access the database.