refactor(model.js): model.js returns Secrets table and has a db client parameter
refactor model.js to return a Secrets table definition and to have a dbClient parameter.

The dbClient will be passed into it in index.js.

This refactor was done to have non-admin user be able to send IAM auth tokens to the embrasure-server so they can connect to the database.

Commented out createTable() function - we are assuming at this point that the Secrets table will already be created.